splitted handlegrip text in serveral modules
