+++ app/scripts/components/handlegrip/handlegrip-bar.components.jsx
@@ -1,0 +1,69 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import PureRenderMixin from 'react-addons-pure-render-mixin';
+import LocalClient from '../../stores/local-client.stores.jsx';
+import Lifespan from 'lifespan';
+import classNames from 'classnames';
+
+/**
+*	Component : the handlegrip (green bar) surrounding the letter
+*	@extends React.Component
+*/
+export default class HandlegripBar extends React.Component {
+	constructor(props) {
+		super(props);
+		this.shouldComponentUpdate = PureRenderMixin.shouldComponentUpdate.bind(this);
+
+		// function bindings
+		this.handleDown = this.handleDown.bind(this);
+	}
+
+	componentWillMount() {
+		this.client = LocalClient.instance();
+		this.lifespan = new Lifespan();
+	}
+
+	componentDidMount() {
+		// retrieve and publicly store the width of the current element
+		this.client.dispatchAction('/store-value', {
+			'uiWordHandleWidth': ReactDOM.findDOMNode(this).offsetWidth,
+		});
+	}
+
+	componentWillUnmount() {
+		this.lifespan.release();
+	}
+
+	handleDown(e) {
+		// tells everyone that the tracking begins, and on which side
+		this.client.dispatchAction('/store-value', {uiSpacingTracking: this.props.side});
+
+		const newX = e.pageX || e.screenX;
+
+		this.client.dispatchAction('/store-value-undoable', {uiTrackingX: newX});
+
+		e.stopPropagation();
+	}
+
+	render() {
+		const left = this.props.side === 'left';
+		const handleGripClasses = classNames({
+			'handlegrip': true,
+			'handlegrip-left': left,
+			'handlegrip-right': !left,
+		});
+
+		return (
+			<span
+				className={handleGripClasses}
+				onMouseDown={this.handleDown}
+				style={this.props.style}
+			>
+				<span className="handlegrip-border"></span>
+				<span className="handlegrip-spacing-number">
+					{this.props.spacing}
+				</span>
+			</span>
+		);
+	}
+}
